Refactor into lib and CLI
This allows Markdown functionality to be incorporated into other C programs.
